Code overhaul, a lot of refactoring
