+++ src/main/java/Main.java
@@ -1,8 +1,8 @@
-import screens.PaintScreen;
+import screens.MainPanel;
 
 public class Main {
     public static void main(String[] args) {
 
-        new PaintScreen();
+        new MainPanel();
     }
 }
